Rename Footer to PageFooter for pages router
