+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import AuthModal from "@/components/AuthModal";
-import Footer from "@/components/Footer";
+import Footer from "@/components/PageFooter";
 import Header from "@/components/PageHeader";
 import "@/styles/globals.css";
 import { api } from "@/utils/api";
